Split Replicator transformer for readability
